Split babel config on device type, rather than modern/legacy browser version
