+++ packages/babel-preset-udemy-website/index.js
@@ -5,10 +5,11 @@ const {declare} = require('@babel/helper-plugin-utils');
 const plugins = require('./plugins');
 
 module.exports = declare((api, options) => {
-    const modernBrowsers =
-        'last 2 Chrome versions, last 2 Firefox versions, last 1 Safari version, last 1 Opera version, last 1 ChromeAndroid version, last 1 FirefoxAndroid version, last 3 iOS versions, last 1 OperaMobile version';
-    const legacyBrowsers = '> 0.2%, last 1 version, not dead';
-    const browsers = options.environment === 'modern' ? modernBrowsers : legacyBrowsers;
+    const mobileBrowsers =
+        'last 2 ChromeAndroid version, last 1 FirefoxAndroid version, last 4 iOS versions, last 1 OperaMobile version';
+    const desktopBrowsers =
+        'last 2 Edge versions, last 5 Chrome versions, last 2 Firefox versions, IE 11, last 1 Opera version, last 2 Safari versions';
+    const browsers = options.device === 'mobile' ? mobileBrowsers : desktopBrowsers;
 
     return {
         presets: [
