+++ packages/babel-preset-udemy-website/tests.js
@@ -41,21 +41,21 @@ process.argv.forEach(val => {
 // eslint-disable-next-line no-console
 console.log('test babel-preset-udemy-website');
 
-const environments = ['legacy', 'modern'];
-const testEnvironment = environment => {
+const devices = ['desktop', 'mobile'];
+const testDevice = device => {
     fs.readdirSync(join(__dirname, 'tests/source'))
         .filter(name => name.endsWith('.js'))
         .forEach(name => {
             const result = babel.transformFileSync(
                 join(__dirname, 'tests/source', name),
-                preset(babel, {environment}),
+                preset(babel, {device}),
             ).code;
-            const expected = fs.readFileSync(join(__dirname, 'tests/result', environment, name), {
+            const expected = fs.readFileSync(join(__dirname, 'tests/result', device, name), {
                 encoding: 'UTF-8',
             });
 
             // eslint-disable-next-line no-console
-            console.log(`\tcheck ${name} -- ${environment}`);
+            console.log(`\tcheck ${name} -- ${device}`);
 
             if (verbose && result.trim() !== expected.trim()) {
                 /* eslint-disable no-console */
@@ -69,17 +69,13 @@ const testEnvironment = environment => {
             if (mode === MODE.TEST) {
                 assert.equal(result.trim(), expected.trim());
             } else if (mode === MODE.UPDATE) {
-                fs.writeFileSync(
-                    join(__dirname, 'tests/result', environment, name),
-                    `${result}\n`,
-                    {
-                        encoding: 'UTF-8',
-                    },
-                );
+                fs.writeFileSync(join(__dirname, 'tests/result', device, name), `${result}\n`, {
+                    encoding: 'UTF-8',
+                });
             }
         });
 };
 
-environments.forEach(e => {
-    testEnvironment(e);
+devices.forEach(e => {
+    testDevice(e);
 });
