+++ packages/babel-preset-udemy-website/tests/result/desktop/css-modules.js
@@ -8,9 +8,7 @@ import "core-js/modules/es6.array.from";
 import "core-js/modules/es7.array.includes";
 import "core-js/modules/es6.array.iterator";
 import "core-js/modules/es6.array.of";
-import "core-js/modules/es6.array.sort";
 import "core-js/modules/es6.array.species";
-import "core-js/modules/es6.date.to-json";
 import "core-js/modules/es6.date.to-primitive";
 import "core-js/modules/es6.function.has-instance";
 import "core-js/modules/es6.function.name";
@@ -60,7 +58,6 @@ import "core-js/modules/es6.object.is-sealed";
 import "core-js/modules/es6.object.is-extensible";
 import "core-js/modules/es6.object.keys";
 import "core-js/modules/es6.object.seal";
-import "core-js/modules/es6.object.set-prototype-of";
 import "core-js/modules/es7.object.values";
 import "core-js/modules/es6.promise";
 import "core-js/modules/es7.promise.finally";
@@ -126,15 +123,41 @@ import "core-js/modules/web.timers";
 import "core-js/modules/web.immediate";
 import "core-js/modules/web.dom.iterable";
 import "regenerator-runtime/runtime";
+import _getClassName from "babel-plugin-react-css-modules/dist/browser/getClassName";
+import React, { Component } from 'react';
+import baseStyles from './css-modules.global.less';
+import styles from './css-modules.less';
+var _styleModuleImportMap = {
+  "styles": {
+    "greeting": "css-modules--greeting--3-JKt"
+  }
+};
 
-function _templateObject() {
-  var data = babelHelpers.taggedTemplateLiteral(["1", "3"]);
+var CssModules =
+/*#__PURE__*/
+function (_Component) {
+  babelHelpers.inherits(CssModules, _Component);
 
-  _templateObject = function _templateObject() {
-    return data;
-  };
+  function CssModules() {
+    babelHelpers.classCallCheck(this, CssModules);
+    return babelHelpers.possibleConstructorReturn(this, babelHelpers.getPrototypeOf(CssModules).apply(this, arguments));
+  }
 
-  return data;
-}
+  babelHelpers.createClass(CssModules, [{
+    key: "render",
+    value: function render() {
+      // styles.greeting can be evaluated at compile time.
+      // baseStyles.colorful won't actually work at runtime since the *.global.less file is
+      // ignored, but I'm testing that it doesn't try to actually output a lookup table
+      // for baseStyles as it will for styles since I'm doing a dynamic styleName.
+      return React.createElement("div", {
+        className: "colorful css-modules--greeting--3-JKt"
+      }, React.createElement("span", {
+        className: _getClassName('baseStyles.colorful', _styleModuleImportMap)
+      }, "Hello World!"));
+    }
+  }]);
+  return CssModules;
+}(Component);
 
-export default String.raw(_templateObject(), 2);
+export { CssModules as default };
