+++ packages/babel-preset-udemy-website/tests/result/mobile/arrows.js
