+++ packages/babel-preset-udemy-website/tests/result/mobile/dynamic-import.js
