+++ packages/babel-preset-udemy-website/tests/result/mobile/for-of-destructuring.js
