+++ packages/babel-preset-udemy-website/tests/result/mobile/object-spread.js
